requirements.txt
numpy>=1.26.0
PyYAML>=6.0.1
pytest>=8.0.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
